+++ lib/src/inline_syntaxes/autolink_extension_syntax.dart
@@ -75,7 +75,7 @@ class AutolinkExtensionSyntax extends InlineSyntax {
     }
 
     return InlineElement(
-      'link',
+      'autolinkExtension',
       children: [Text.fromSpan(span)],
       attributes: {'destination': destination},
     );